requirements.txt
attr
click
delegator-py
enum34 ; python_version < '3.4'
future
inflection
mako
packaging
path.py
requirementslib
six
tomlkit
typing ; python_version < '3.6'
